improving sorting script to merge directories

+++ .config/zsh/sort.sh
@@ -1,11 +1,33 @@
 #! /bin/bash
-cp -fr /run/media/irfan/Programming/cp/At* /run/media/irfan/Programming/cp/atcoder
-rm -fr /run/media/irfan/Programming/cp/At* 
-cp -fr /run/media/irfan/Programming/cp/Ha* /run/media/irfan/Programming/cp/hackerrank
-rm -fr /run/media/irfan/Programming/cp/Ha* 
-cp -fr /run/media/irfan/Programming/cp/Codef* /run/media/irfan/Programming/cp/codeforces
-rm -fr /run/media/irfan/Programming/cp/Codef*
-cp -fr /run/media/irfan/Programming/cp/CodeShef* /run/media/irfan/Programming/cp/codechef
-rm -fr /run/media/irfan/Programming/cp/CodeShef*  
+mv /run/media/irfan/Programming/cp/Codeforc* /run/media/irfan/Programming/cp/codeforces || command_failed=1
+if [ ${command_failed} -eq 1 ]
+then
+  cp -fr /run/media/irfan/Programming/cp/Codef* /run/media/irfan/Programming/cp/codeforces
+  rm -fr /run/media/irfan/Programming/cp/Codef*
+fi
+
+
+# mv /run/media/irfan/Programming/cp/Ha* /run/media/irfan/Programming/cp/hackerrank/ || cmmand_failed=1
+# if [ ${cmmand_failed} -eq 1 ]
+# then
+#   cp -fr /run/media/irfan/Programming/cp/Ha* /run/media/irfan/Programming/cp/hackerrank
+#   rm -fr /run/media/irfan/Programming/cp/Ha* 
+# fi
+
+# mv /run/media/irfan/Programming/cp/At* /run/media/irfan/Programming/cp/atcoder/ || cmand_failed=1
+# if [ ${cmand_failed} -eq 1 ]
+# then
+# cp -fr /run/media/irfan/Programming/cp/At* /run/media/irfan/Programming/cp/atcoder
+# rm -fr /run/media/irfan/Programming/cp/At* 
+# fi
+
+
+# mv /run/media/irfan/Programming/cp/CodeShef* /run/media/irfan/Programming/cp/codechef/ || cmnd_failed=1
+# if [ ${cmnd_failed} -eq 1 ]
+# then
+# cp -fr /run/media/irfan/Programming/cp/CodeShef* /run/media/irfan/Programming/cp/codechef
+# rm -fr /run/media/irfan/Programming/cp/CodeShef*  
+# fi
+
 rm -fr /run/media/irfan/Programming/cp/*.exe
 clear
